eng/requirements.txt
paho-mqtt>=1.5.1
pymongo>=3.11.4
python-socketio>=5.3.0
requests>=2.25.1
